+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-CMXNB2M2.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-BXZ4XYyR.js";import{_ as a}from"./mermaid.core-Bem2b15F.js";import"./chunk-5VM5RSS4-_Bqmuqgj.js";import"./chunk-XXDRQBXY-TghFxgzz.js";import"./chunk-POPQ4Y6H-CnAVQVgT.js";import"./chunk-F27PBJKO-BFEziCqz.js";import"./channel-CfFh7ru6.js";import"./index-DBUbiQL7.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};